* commit '82ad9e320aab49a25b66d01cba4e78e673cd5f6f':
  use longs instead of ints to store pointers in OpenSSL binding

+++ luni/src/main/java/org/apache/harmony/xnet/provider/jsse/AbstractSessionContext.java
@@ -41,7 +41,7 @@ abstract class AbstractSessionContext implements SSLSessionContext {
     volatile int maximumSize;
     volatile int timeout;
 
-    final int sslCtxNativePointer = NativeCrypto.SSL_CTX_new();
+    final long sslCtxNativePointer = NativeCrypto.SSL_CTX_new();
 
     /** Identifies OpenSSL sessions. */
     static final int OPEN_SSL = 1;
