+++ luni/src/main/java/org/apache/harmony/xnet/provider/jsse/OpenSSLDigestContext.java
@@ -17,9 +17,9 @@
 package org.apache.harmony.xnet.provider.jsse;
 
 public class OpenSSLDigestContext {
-    private final int context;
+    private final long context;
 
-    public OpenSSLDigestContext(int ctx) {
+    public OpenSSLDigestContext(long ctx) {
         if (ctx == 0) {
             throw new NullPointerException("ctx == 0");
         }
@@ -36,7 +36,7 @@ public class OpenSSLDigestContext {
         }
     }
 
-    int getContext() {
+    long getContext() {
         return context;
     }
 }
